Consolidate Platform.cpp
